Change column ci to bigint

+++ Tables/dbo.Test.sql
@@ -1,6 +1,6 @@
 CREATE TABLE [dbo].[Test]
 (
-[c1] [int] NULL,
+[c1] [bigint] NULL,
 [c2] [nvarchar] (50) COLLATE SQL_Latin1_General_CP1_CI_AS NULL
 ) ON [PRIMARY]
 GO
